Add pytest conftest ...
