+++ tests/slickml/classification/test_glmnet.py
@@ -7,53 +7,15 @@ import pytest
 import shap
 from assertpy import assert_that
 from matplotlib.figure import Figure
-from pytest import FixtureRequest
-from sklearn.model_selection import train_test_split
 
 from slickml.classification import GLMNetCVClassifier
-from tests.utils import _ids, _load_test_data_from_csv
+from tests.conftest import _ids
 
 
 # TODO(amir): add lolipop plot for coeff + unit-test
 # TODO(amir): add test for lambda-path param
 class TestGLMNetCVClassifier:
     """Validates `GLMNetCVClassifier` instantiation."""
-
-    @staticmethod
-    @pytest.fixture(scope="module")
-    def clf_x_y_data(
-        request: FixtureRequest,
-    ) -> Tuple[
-        Union[pd.DataFrame, np.ndarray],
-        Union[pd.DataFrame, np.ndarray],
-        Union[np.ndarray, List],
-        Union[np.ndarray, List],
-    ]:
-        """Returns stratified train/test sets."""
-        df = _load_test_data_from_csv(
-            filename="clf_test_data.csv",
-        )
-        y = df["CLASS"].values
-        X = df.drop(
-            ["CLASS"],
-            axis=1,
-        )
-        X_train, X_test, y_train, y_test = train_test_split(
-            X,
-            y,
-            test_size=0.2,
-            shuffle=True,
-            stratify=y,
-            random_state=1367,
-        )
-        if request.param == "dataframe":
-            return (X_train, X_test, y_train, y_test)
-        elif request.param == "array":
-            return (X_train.values, X_test.values, y_train, y_test)
-        elif request.param == "list":
-            return (X_train, X_test, y_train.tolist(), y_test.tolist())
-        else:
-            return None
 
     @pytest.mark.parametrize(
         ("kwargs"),
@@ -76,18 +38,18 @@ class TestGLMNetCVClassifier:
             GLMNetCVClassifier(**kwargs)
 
     @pytest.mark.parametrize(
-        ("clf_x_y_data"),
+        ("clf_train_test_x_y"),
         [
             ("array"),
             ("dataframe"),
             ("list"),
         ],
-        indirect=["clf_x_y_data"],
+        indirect=["clf_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvclassifier__passes__with_defaults_and_no_test_targets(
         self,
-        clf_x_y_data: Tuple[
+        clf_train_test_x_y: Tuple[
             Union[pd.DataFrame, np.ndarray],
             Union[pd.DataFrame, np.ndarray],
             Union[np.ndarray, List],
@@ -95,7 +57,7 @@ class TestGLMNetCVClassifier:
         ],
     ) -> None:
         """Validates `GLMNetCVClassifier` instanation passes with default inputs."""
-        X_train, X_test, y_train, _ = clf_x_y_data
+        X_train, X_test, y_train, _ = clf_train_test_x_y
         clf = GLMNetCVClassifier()
         clf.fit(X_train, y_train)
         y_pred_proba = clf.predict_proba(X_test)
@@ -184,18 +146,18 @@ class TestGLMNetCVClassifier:
         npt.assert_almost_equal(np.mean(clf.shap_values_train_), 0.01112, decimal=5)
 
     @pytest.mark.parametrize(
-        ("clf_x_y_data"),
+        ("clf_train_test_x_y"),
         [
             ("array"),
             ("dataframe"),
             ("list"),
         ],
-        indirect=["clf_x_y_data"],
+        indirect=["clf_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvclassifier__passes__with_defaults(
         self,
-        clf_x_y_data: Tuple[
+        clf_train_test_x_y: Tuple[
             Union[pd.DataFrame, np.ndarray],
             Union[pd.DataFrame, np.ndarray],
             Union[np.ndarray, List],
@@ -203,7 +165,7 @@ class TestGLMNetCVClassifier:
         ],
     ) -> None:
         """Validates `GLMNetCVClassifier` instanation passes with default inputs."""
-        X_train, X_test, y_train, y_test = clf_x_y_data
+        X_train, X_test, y_train, y_test = clf_train_test_x_y
         clf = GLMNetCVClassifier()
         clf.fit(X_train, y_train)
         # Note: we pass `y_test` for the sake of testing while in inference we might night have
@@ -296,7 +258,7 @@ class TestGLMNetCVClassifier:
 
     # TODO(amir): add a test for `lambda_path` parameter
     @pytest.mark.parametrize(
-        ("clf_x_y_data", "kwargs"),
+        ("clf_train_test_x_y", "kwargs"),
         [
             ("dataframe", {"alpha": 0.9}),
             ("dataframe", {"n_lambda": 200}),
@@ -308,12 +270,12 @@ class TestGLMNetCVClassifier:
             ("dataframe", {"random_state": 42}),
             ("dataframe", {"max_features": 10}),
         ],
-        indirect=["clf_x_y_data"],
+        indirect=["clf_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvclassifier__passes__with_valid_inputs(
         self,
-        clf_x_y_data: Tuple[
+        clf_train_test_x_y: Tuple[
             Union[pd.DataFrame, np.ndarray],
             Union[pd.DataFrame, np.ndarray],
             Union[np.ndarray, List],
@@ -322,7 +284,7 @@ class TestGLMNetCVClassifier:
         kwargs: Optional[Dict[str, Any]],
     ) -> None:
         """Validates `GLMNetCVClassifier` instanation passes with valid inputs."""
-        X_train, X_test, y_train, y_test = clf_x_y_data
+        X_train, X_test, y_train, y_test = clf_train_test_x_y
         clf = GLMNetCVClassifier(**kwargs)
         clf.fit(X_train, y_train)
         # Note: we pass `y_test` for the sake of testing while in inference we might night have
@@ -404,7 +366,7 @@ class TestGLMNetCVClassifier:
 
     @pytest.mark.parametrize(
         (
-            "clf_x_y_data",
+            "clf_train_test_x_y",
             "waterfall_kwargs",
             "summary_kwargs",
         ),
@@ -450,17 +412,17 @@ class TestGLMNetCVClassifier:
                 },
             ),
         ],
-        indirect=["clf_x_y_data"],
+        indirect=["clf_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvclassifier_shap_plots__passes__with_valid_inputs(
         self,
-        clf_x_y_data: Tuple[pd.DataFrame, pd.DataFrame, np.ndarray, np.ndarray],
+        clf_train_test_x_y: Tuple[pd.DataFrame, pd.DataFrame, np.ndarray, np.ndarray],
         waterfall_kwargs: Dict[str, Any],
         summary_kwargs: Dict[str, Any],
     ) -> None:
         """Validates `GLMNetCVClassifier` Shap plots passes with valid inputs."""
-        X_train, X_test, y_train, y_test = clf_x_y_data
+        X_train, X_test, y_train, y_test = clf_train_test_x_y
         clf = GLMNetCVClassifier()
         clf.fit(X_train, y_train)
         _ = clf.predict_proba(X_test, y_test)
