+++ tests/slickml/regression/test_glmnet.py
@@ -7,11 +7,9 @@ import pytest
 import shap
 from assertpy import assert_that
 from matplotlib.figure import Figure
-from pytest import FixtureRequest
-from sklearn.model_selection import train_test_split
 
 from slickml.regression import GLMNetCVRegressor
-from tests.utils import _ids, _load_test_data_from_csv
+from tests.conftest import _ids
 
 
 # TODO(amir): add lolipop plot for coeff + unit-test
@@ -19,42 +17,6 @@ from tests.utils import _ids, _load_test_data_from_csv
 # TODO(amir): test out regression with multiple outputs (y1, y2)
 class TestGLMNetCVRegressor:
     """Validates `GLMNetCVRegressor` instantiation."""
-
-    @staticmethod
-    @pytest.fixture(scope="module")
-    def reg_x_y_data(
-        request: FixtureRequest,
-    ) -> Tuple[
-        Union[pd.DataFrame, np.ndarray],
-        Union[pd.DataFrame, np.ndarray],
-        Union[np.ndarray, List],
-        Union[np.ndarray, List],
-    ]:
-        """Returns train/test sets."""
-        df = _load_test_data_from_csv(
-            filename="reg_test_data.csv",
-        )
-        # TODO(amir): try to pull-out multi target regression as well here
-        y = df["TARGET1"].values
-        X = df.drop(
-            ["TARGET1", "TARGET2"],
-            axis=1,
-        )
-        X_train, X_test, y_train, y_test = train_test_split(
-            X,
-            y,
-            test_size=0.2,
-            shuffle=True,
-            random_state=1367,
-        )
-        if request.param == "dataframe":
-            return (X_train, X_test, y_train, y_test)
-        elif request.param == "array":
-            return (X_train.values, X_test.values, y_train, y_test)
-        elif request.param == "list":
-            return (X_train, X_test, y_train.tolist(), y_test.tolist())
-        else:
-            return None
 
     @pytest.mark.parametrize(
         ("kwargs"),
@@ -77,18 +39,18 @@ class TestGLMNetCVRegressor:
             GLMNetCVRegressor(**kwargs)
 
     @pytest.mark.parametrize(
-        ("reg_x_y_data"),
+        ("reg_train_test_x_y"),
         [
             ("array"),
             ("dataframe"),
             ("list"),
         ],
-        indirect=["reg_x_y_data"],
+        indirect=["reg_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvregressor__passes__with_defaults_and_no_test_targets(
         self,
-        reg_x_y_data: Tuple[
+        reg_train_test_x_y: Tuple[
             Union[pd.DataFrame, np.ndarray],
             Union[pd.DataFrame, np.ndarray],
             Union[np.ndarray, List],
@@ -96,7 +58,7 @@ class TestGLMNetCVRegressor:
         ],
     ) -> None:
         """Validates `GLMNetCVRegressor` instanation passes with default inputs."""
-        X_train, X_test, y_train, _ = reg_x_y_data
+        X_train, X_test, y_train, _ = reg_train_test_x_y
         reg = GLMNetCVRegressor()
         reg.fit(X_train, y_train)
         y_pred = reg.predict(X_test)
@@ -182,18 +144,18 @@ class TestGLMNetCVRegressor:
         npt.assert_almost_equal(np.mean(reg.shap_values_train_), 8.13e-06, decimal=5)
 
     @pytest.mark.parametrize(
-        ("reg_x_y_data"),
+        ("reg_train_test_x_y"),
         [
             ("array"),
             ("dataframe"),
             ("list"),
         ],
-        indirect=["reg_x_y_data"],
+        indirect=["reg_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvregressor__passes__with_defaults(
         self,
-        reg_x_y_data: Tuple[
+        reg_train_test_x_y: Tuple[
             Union[pd.DataFrame, np.ndarray],
             Union[pd.DataFrame, np.ndarray],
             Union[np.ndarray, List],
@@ -201,7 +163,7 @@ class TestGLMNetCVRegressor:
         ],
     ) -> None:
         """Validates `GLMNetCVRegressor` instanation passes with default inputs."""
-        X_train, X_test, y_train, y_test = reg_x_y_data
+        X_train, X_test, y_train, y_test = reg_train_test_x_y
         reg = GLMNetCVRegressor()
         reg.fit(X_train, y_train)
         # Note: we pass `y_test` for the sake of testing while in inference we might night have
@@ -291,11 +253,11 @@ class TestGLMNetCVRegressor:
 
     # TODO(amir): add a test for `lambda_path` parameter
     @pytest.mark.parametrize(
-        ("reg_x_y_data", "kwargs"),
+        ("reg_train_test_x_y", "kwargs"),
         [
-            ("dataframe", {"alpha": 0.9}),
-            ("dataframe", {"n_lambda": 200}),
-            ("dataframe", {"n_splits": 10}),
+            ("dataframe", {"alpha": 0.1}),
+            ("dataframe", {"n_lambda": 50}),
+            ("dataframe", {"n_splits": 5}),
             ("dataframe", {"metric": "mean_squared_error"}),
             ("dataframe", {"scale": False, "sparse_matrix": True}),
             ("dataframe", {"fit_intercept": False}),
@@ -303,12 +265,12 @@ class TestGLMNetCVRegressor:
             ("dataframe", {"random_state": 42}),
             ("dataframe", {"max_features": 10}),
         ],
-        indirect=["reg_x_y_data"],
+        indirect=["reg_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvregressor__passes__with_valid_inputs(
         self,
-        reg_x_y_data: Tuple[
+        reg_train_test_x_y: Tuple[
             Union[pd.DataFrame, np.ndarray],
             Union[pd.DataFrame, np.ndarray],
             Union[np.ndarray, List],
@@ -317,7 +279,7 @@ class TestGLMNetCVRegressor:
         kwargs: Optional[Dict[str, Any]],
     ) -> None:
         """Validates `GLMNetCVRegressor` instanation passes with valid inputs."""
-        X_train, X_test, y_train, y_test = reg_x_y_data
+        X_train, X_test, y_train, y_test = reg_train_test_x_y
         reg = GLMNetCVRegressor(**kwargs)
         reg.fit(X_train, y_train)
         # Note: we pass `y_test` for the sake of testing while in inference we might night have
@@ -397,7 +359,7 @@ class TestGLMNetCVRegressor:
 
     @pytest.mark.parametrize(
         (
-            "reg_x_y_data",
+            "reg_train_test_x_y",
             "waterfall_kwargs",
             "summary_kwargs",
         ),
@@ -442,17 +404,17 @@ class TestGLMNetCVRegressor:
                 },
             ),
         ],
-        indirect=["reg_x_y_data"],
+        indirect=["reg_train_test_x_y"],
         ids=_ids,
     )
     def test_glmnetcvregressor_shap_plots__passes__with_valid_inputs(
         self,
-        reg_x_y_data: Tuple[pd.DataFrame, pd.DataFrame, np.ndarray, np.ndarray],
+        reg_train_test_x_y: Tuple[pd.DataFrame, pd.DataFrame, np.ndarray, np.ndarray],
         waterfall_kwargs: Dict[str, Any],
         summary_kwargs: Dict[str, Any],
     ) -> None:
         """Validates `GLMNetCVRegressor` Shap plots passes with valid inputs."""
-        X_train, X_test, y_train, y_test = reg_x_y_data
+        X_train, X_test, y_train, y_test = reg_train_test_x_y
         reg = GLMNetCVRegressor()
         reg.fit(X_train, y_train)
         _ = reg.predict(X_test, y_test)
